Unite two test files related to external cads
